注册登录重构:昵称+邮箱验证码+密码注册,账号ID=int自增起始10000(含PG迁移),登录仅账号+密码,好友按账号ID添加;冒烟测试16/16

+++ LHZ.OnlineChat.Server/Models/DTOs/MessageDtos.cs
@@ -6,7 +6,7 @@ namespace LHZ.OnlineChat.Server.Models.DTOs;
 public class MessageDto
 {
     public long Id { get; set; }
-    public long SenderId { get; set; }
+    public int SenderId { get; set; }
     public string SenderName { get; set; } = string.Empty;
     public string? SenderAvatar { get; set; }
     public string Content { get; set; } = string.Empty;
@@ -19,7 +19,7 @@ public class GroupMessageDto
 {
     public long Id { get; set; }
     public long GroupId { get; set; }
-    public long SenderId { get; set; }
+    public int SenderId { get; set; }
     public string SenderName { get; set; } = string.Empty;
     public string? SenderAvatar { get; set; }
     public string Content { get; set; } = string.Empty;
