+++ LHZ.OnlineChat.Server/Models/Entities/PrivateMessage.cs
@@ -13,10 +13,10 @@ public class PrivateMessage
     public long Id { get; set; }
 
     [Column(IsNullable = false)]
-    public long SenderId { get; set; }
+    public int SenderId { get; set; }
 
     [Column(IsNullable = false)]
-    public long ReceiverId { get; set; }
+    public int ReceiverId { get; set; }
 
     [Column(IsNullable = false, StringLength = -2)]
     public string Content { get; set; } = string.Empty;
